flask/pyproject.toml
[tool.poetry]
name = "hello-oidc-auth-with-oauth2-proxy-and-fastapi"
version = "0.1.0"
description = ""
authors = ["Markus Katharina Brechtel <markus.brechtel@uk-koeln.de>"]
license = "Apache2"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
